CHG: [C++] Move out general functionality to common
